Error  in Mean value calculation
In line 36 the histc function doesn't give the number of  occurences for each unique element. Histc measures the number of rows that are included in the bins specified from Get_unique_rows variable. In that way we do not get the number of discrete occurences that are affiliated with the specific rows, but the number of rows that are included in the bin.